+++ src/anomalib/data/datamodules/__init__.py
@@ -4,7 +4,19 @@
 """Anomalib Data Modules."""
 
 from .depth import Folder3D, MVTec3D
-from .image import BMAD, MPDD, VAD, BTech, Datumaro, Folder, Kolektor, MVTecAD, Tabular, Visa
+from .image import (
+    BMAD,
+    MPDD,
+    VAD,
+    AutoVI,
+    BTech,
+    Datumaro,
+    Folder,
+    Kolektor,
+    MVTecAD,
+    Tabular,
+    Visa,
+)
 from .video import Avenue, ShanghaiTech, UCSDped
 
 __all__ = [
@@ -23,4 +35,5 @@ __all__ = [
     "ShanghaiTech",
     "UCSDped",
     "BMAD",
+    "AutoVI",
 ]